doc: fix undici version in changelog
The last PR to update the undici version in V18 and
V16 had two commits. Once to update to 5.9.1 and one
to upodate to 5.10.0. Unfortunately it looks like
when squashed the commit indicates 5.9.1 which was
reported as confusing in the changelogs.

I don't think we can necessarily change the commit
message in the commit, but I think having the
changelogs indicate the correct version makes
sense even without that change.

This PR updates the version in the two changelogs.

Signed-off-by: Michael Dawson <mdawson@devrus.com>

PR-URL: https://github.com/nodejs/node/pull/44982
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>
Reviewed-By: Beth Griggs <bethanyngriggs@gmail.com>
Reviewed-By: Yagiz Nizipli <yagiz@nizipli.com>